TrainAFINE/requirements.txt
addict
future
lmdb
numpy==1.23.0
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tensorboard
tqdm
yapf
torchprofile
pytorch-fid
lpips
matplotlib
jupyter
wandb
timm
sk-video>=1.1.10
moviepy>=1.0.3
einops
ftfy
regex
dists-pytorch
loguru
prettytable
ninja
nvidia-ml-py3
fvcore
pytorch-msssim
openpyxl
